change the cvp meaning so as to be unique for every web service
During bridging pass the cpp and the selections so as to get the desired xslt

Register a new web service as a wsdl file

Add showServices.jsp

Add serviceReg.jsp

Add WSDL parser

Get all services that belong to a software component. if cvp is true the functions only returns those that have been annotated. We put that boolean condition so as to hide services that have never been annotated from organization

create the CVP entity at the first intent of annotation by the Vendor  (at a functional or data level)
create the CPP entity at the first intent of annotation by the Organization (at a functional or data level)